+++ hutool-core/src/main/java/com/xiaoleilu/hutool/io/IoUtil.java
@@ -484,7 +484,7 @@ public final class IoUtil {
 		} catch (IOException e) {
 			throw new IORuntimeException(e);
 		}
-		if (readLength < length) {
+		if (readLength > 0 && readLength < length) {
 			byte[] b2 = new byte[length];
 			System.arraycopy(b, 0, b2, 0, readLength);
 			return b2;