+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/ReferenceUtil.java
@@ -27,8 +27,8 @@ public class ReferenceUtil {
 	 * @param referent 被引用对象
 	 * @return {@link Reference}
 	 */
-	public static <T> Reference<T> createReference(ReferenceType type, T referent) {
-		return createReference(type, referent, null);
+	public static <T> Reference<T> create(ReferenceType type, T referent) {
+		return create(type, referent, null);
 	}
 
 	/**
@@ -39,7 +39,7 @@ public class ReferenceUtil {
 	 * @param queue 引用队列
 	 * @return {@link Reference}
 	 */
-	public static <T> Reference<T> createReference(ReferenceType type, T referent, ReferenceQueue<T> queue) {
+	public static <T> Reference<T> create(ReferenceType type, T referent, ReferenceQueue<T> queue) {
 		switch (type) {
 		case SOFT:
 			return new WeakReference<>(referent);