+++ hutool-core/src/test/java/com/xiaoleilu/hutool/core/io/FileTypeUtilTest.java
@@ -1,13 +1,13 @@
 package com.xiaoleilu.hutool.core.io;
 
 import java.io.File;
-import java.io.IOException;
 
 import org.junit.Assert;
 import org.junit.Test;
 
 import com.xiaoleilu.hutool.io.FileTypeUtil;
 import com.xiaoleilu.hutool.io.FileUtil;
+import com.xiaoleilu.hutool.lang.Console;
 
 /**
  * 文件类型判断单元测试
@@ -17,7 +17,7 @@ import com.xiaoleilu.hutool.io.FileUtil;
 public class FileTypeUtilTest {
 	
 	@Test
-	public void fileTypeUtilTest() throws IOException{
+	public void fileTypeUtilTest() {
 		File file = FileUtil.file("hutool.jpg");
 		String type = FileTypeUtil.getType(file);
 		Assert.assertEquals("jpg", type);
@@ -25,5 +25,12 @@ public class FileTypeUtilTest {
 		FileTypeUtil.putFileType("ffd8ffe000104a464946", "new_jpg");
 		String newType = FileTypeUtil.getType(file);
 		Assert.assertEquals("new_jpg", newType);
+	}
+	
+	@Test
+	public void emptyTest() {
+		File file = FileUtil.file("d:/empty.txt");
+		String type = FileTypeUtil.getType(file);
+		Console.log(type);
 	}
 }